% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dfm.R
\name{dfm}
\alias{dfm}
\title{Estimates a dynamic factor model based on Doz, Gianone & Reichlin (2011)}
\usage{
dfm(X, r, p, q, max_iter = 100, threshold = 1e-04, rQ, rC)
}
\arguments{
\item{X}{Data matrix (T x n)}

\item{r}{Number of static factors}

\item{p}{Lag order for factors}

\item{q}{Number of dynamic factors, must be equal to or less than r}

\item{max_iter}{Maximum number of iterations in EM-algorithm}

\item{threshold}{Threshold for algorithm convergence}

\item{rQ}{Restrictions on system state covariance}

\item{rC}{Restrictions on factor loading matrix}
}
\value{
3 types of factor estimates, namely principal component estimate, two step
estimate based on PCA and Kalman filtering and QML estimate based on EM-algorithm
}
\description{
Estimates a dynamic factor model based on Doz, Gianone & Reichlin (2011)
}
\examples{
x <- matrix(rnorm(50*10), 50, 10)
W <- as.logical(matrix(rbinom(50*10, 1, 0.1), 50, 10))
x[W] <- NA
dfm(x, 2, 2, 1)
}

