% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dfmMS.R
\name{dfmMS}
\alias{dfmMS}
\title{Dynamic factor model with Markov-switching states}
\usage{
dfmMS(X, J = 1, s = 2, p = 1, x0, P0)
}
\arguments{
\item{X}{Data matrix (Txn)}

\item{J}{Number of factors. Currently, only \code{1} is supported}

\item{s}{Number of states. Only \code{2} are supported for now}

\item{p}{Lag number for VAR model of factors}

\item{x0}{Initial value for state}

\item{P0}{Initial value for state covariance}
}
\value{
Return an estimator. Currently, it runs via a likelihood
maximization an so is rather slow. Do not call it with large number
of input variables or missing data. Some box constraints are currently
to speed up the process, these are based on usual rationale for economic
data.
}
\description{
Dynamic factor model with Markov-switching states
}

