% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kalman.R
\name{K_filter}
\alias{K_filter}
\title{Implements a Kalman for dynamic factor model.}
\usage{
K_filter(initx, initV, x, A, C, R, Q)
}
\arguments{
\item{initx}{Initial value for state space observations}

\item{initV}{Initial value for state covariance}

\item{x}{Observation matrix}

\item{A}{State space matrix}

\item{C}{System matrix}

\item{R}{State space covariance}

\item{Q}{System covariance}
}
\value{
Filtered state space variable and its covariance matrix
as well as their forecast for next period for further iterations
}
\description{
Implements a Kalman for dynamic factor model.
}

