% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kalman.R
\name{K_smoother}
\alias{K_smoother}
\title{Implements Kalman smoothing and is used along with Kalman filter.
Kalman filter outputs enter Kalman smoother as inputs.}
\usage{
K_smoother(A, xitt, xittm, Ptt, Pttm, C, R, W)
}
\arguments{
\item{A}{State space matrix}

\item{xitt}{State space variable}

\item{xittm}{Predicted state space variable}

\item{Ptt}{State space covariance}

\item{Pttm}{Predicted state space covariance}

\item{C}{System matrix}

\item{R}{State space covariance}

\item{W}{Logical matrix (T x n) indicating missing data.
TRUE if observation is present, FALSE if it is missing.}
}
\value{
Smoothed state space variable and state space covariance matrix
}
\description{
Implements Kalman smoothing and is used along with Kalman filter.
Kalman filter outputs enter Kalman smoother as inputs.
}

