% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kim.R
\name{KimFilter}
\alias{KimFilter}
\title{Implementation of Kim (1994) filter, an extension to Kalman filter
for dynamic linear models with Markov-switching. Documentation
is incomplete, rudimentary and needs to be rechecked!}
\usage{
KimFilter(x0, P0, y, F, A, R, Q, p)
}
\arguments{
\item{x0}{Initial condition for state vector}

\item{P0}{Initial condition for state variance}

\item{y}{Data matrix (Txn)}

\item{F}{System matrix for measurement equation}

\item{A}{Transition matrix for state equation}

\item{R}{Error covariance for measurement equation}

\item{Q}{Error covariance for state equation}

\item{p}{Transition probability matrix}
}
\value{
Filtered states and covariances with associated probability matrices.
}
\description{
Implementation of Kim (1994) filter, an extension to Kalman filter
for dynamic linear models with Markov-switching. Documentation
is incomplete, rudimentary and needs to be rechecked!
}

