% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kim.R
\name{KimSmoother}
\alias{KimSmoother}
\title{Smoothing algorithm from Kim (1994) to be used following a run
of KimFilter function.}
\usage{
KimSmoother(xA, Pa, A, P, x, p, stateP, stateP_fut)
}
\arguments{
\item{xA}{Filtered state vector to be smoothed}

\item{Pa}{Filtered state covariance to be smoothed}

\item{A}{Array with transition matrices}

\item{P}{State-dependent state covariance}

\item{x}{State-dependent state vector}

\item{p}{Markov transition matrix}

\item{stateP}{Evolving current probability matrix}

\item{stateP_fut}{Predicted probability matrix}
}
\value{
Smoothed states and covariance matrices. This is the equivalent
of Kalman smoother in Markov-switching case.
}
\description{
Smoothing algorithm from Kim (1994) to be used following a run
of KimFilter function.
}

