% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dfm.R
\docType{data}
\name{NBBsurvey}
\alias{NBBsurvey}
\title{National Bank of Belgium business and consumer surveys}
\format{A time-stamped dataframe (xts object) with}
\source{
\url{http://stat.nbb.be}
}
\usage{
NBBsurvey
}
\description{
National Bank of Belgium business and consumer surveys
}
\keyword{datasets}

