% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.dfm}
\alias{predict.dfm}
\title{Predict factors and observables based on an estimated dynamic
factor model}
\usage{
\method{predict}{dfm}(x, h = 1)
}
\arguments{
\item{x}{An object of \code{dfm} class}

\item{h}{Forecasting horizon}
}
\value{
Returns prediction matrices for factors and input variables
based on model estimation. Each row corresponds to a prediction
horizon. First row is \code{T+1}, second -\code{T+2} and so on...
}
\description{
Predict factors and observables based on an estimated dynamic
factor model
}

