% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/additional.R
\name{VAR}
\alias{VAR}
\title{Estimate a p-th order vector autoregressive (VAR) model}
\usage{
VAR(x, p)
}
\arguments{
\item{x}{Data matrix (T x n)}

\item{p}{Maximum lag order, i.e. VAR(p) will be estimated}
}
\value{
Estimated parameter matrix, residuals and regression model
independent and dependent variables
}
\description{
Estimate a p-th order vector autoregressive (VAR) model
}
\examples{
x <- matrix(rnorm(50*2), nrow=50, ncol=2)
VAR(x, 2)
}

